Signal to speed conversion for the robot
The robot is now a blueprint with the graphics and a cpp class hooked up. The cpp class has the capability to move the robot when you press "G".

Also the call side of the Moving Camera event dispatcher should be set up, but I haven't tried binding it or tested it yet.